+++ Mobile Development/SHARA/app/src/main/java/com/example/shara/ui/history/HistoryFragment.kt
@@ -1,4 +1,4 @@
-package com.example.shara.ui.notifications
+package com.example.shara.ui.history
 
 import android.os.Bundle
 import android.view.LayoutInflater
