+++ Mobile Development/SHARA/app/src/main/res/navigation/mobile_navigation.xml
@@ -19,7 +19,7 @@
 
     <fragment
         android:id="@+id/navigation_history"
-        android:name="com.example.shara.ui.notifications.HistoryFragment"
+        android:name="com.example.shara.ui.history.HistoryFragment"
         android:label="@string/title_history"
         tools:layout="@layout/fragment_notifications" />
 </navigation>
